requirements.txt
Random-Word
aiohttp==3.8.3
requests==2.28.1
lxml
beautifulsoup4==4.11.1
Pillow==9.2.0
asyncio==3.4.3
pyTelegramBotAPI==4.7.1
fake_useragent
aiofile
SQLAlchemy
alembic
aiofiles
fake-useragent
apscheduler
pandas
asincio

